# services/agent/docs/openapi/agent.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Actinver AI Advisor - Agent API",
    "description": "Mobile/web-facing API of the Actinver AI Advisor (bff-mobile + agent-orchestrator + avatar-broker). Path-versioned (`/v1/`), RFC 9457 errors with a Spanish `message`, `Idempotency-Key` on every mutating endpoint, cursor pagination, SSE for chat turns and a WebSocket for the bidirectional audio path. Auth: `Authorization: DPoP <token>` plus a `DPoP` proof header (plain `Bearer` accepted in local only).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/threads/{thread_id}/forms/{form_id}/submit": {
      "post": {
        "tags": [
          "forms"
        ],
        "summary": "Submit a Form Spec with a step-up assertion",
        "description": "Requires `Idempotency-Key`. The server re-verifies the spec signature and TTL, re-enters the suitability gate if the amount changed, verifies the step-up signature against the registered device key and only then places the order. Errors: `FORM_EXPIRED` 409, `FORM_SIGNATURE_INVALID` 400, `FORM_ALREADY_USED` 409, `ACK_REQUIRED` 422, `STEP_UP_REQUIRED` 401, `LIMIT_EXCEEDED` 422, `NO_SUITABLE_PRODUCT` 422.",
        "operationId": "submit_form_v1_threads__thread_id__forms__form_id__submit_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          },
          {
            "name": "form_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Form Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "DPoP",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dpop"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/step-up/challenge": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Issue a single-use step-up challenge",
        "description": "Requires `Idempotency-Key`. The challenge is bound to `form_id` and the declared amount and expires in 120 s. Sign the raw challenge bytes with the device key (ES256, raw r||s) and send the base64url signature as `step_up_assertion`.",
        "operationId": "step_up_challenge_v1_auth_step_up_challenge_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "DPoP",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dpop"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepUpChallengeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepUpChallengeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FormSubmitRequest": {
        "properties": {
          "values": {
            "additionalProperties": true,
            "type": "object",
            "title": "Values"
          },
          "acknowledgements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Acknowledgements"
          },
          "challenge_id": {
            "type": "string",
            "title": "Challenge Id"
          },
          "step_up_assertion": {
            "type": "string",
            "title": "Step Up Assertion",
            "description": "base64url ES256 signature over the challenge"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "values",
          "challenge_id",
          "step_up_assertion"
        ],
        "title": "FormSubmitRequest"
      },
      "FormSubmitResponse": {
        "properties": {
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "settlement_date": {
            "type": "string",
            "title": "Settlement Date"
          },
          "evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Id"
          },
          "ui_payload": {
            "items": {
              "$ref": "#/components/schemas/UIComponent"
            },
            "type": "array",
            "title": "Ui Payload"
          },
          "speech": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speech"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "status",
          "settlement_date",
          "evidence_id"
        ],
        "title": "FormSubmitResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Money": {
        "properties": {
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "enum": [
              "MXN",
              "USD",
              "EUR"
            ],
            "title": "Currency"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount",
          "currency"
        ],
        "title": "Money",
        "description": "Always ``{amount: decimal string, currency}`` - never a bare number\n(docs/04-backend/04 §1). A value without a currency is rejected, not\ndefaulted to MXN (docs/07-data-governance/03 §6)."
      },
...
      "StepUpChallengeRequest": {
        "properties": {
          "form_id": {
            "type": "string",
            "title": "Form Id"
          },
          "amount": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "form_id",
          "amount"
        ],
        "title": "StepUpChallengeRequest"
      },
      "StepUpChallengeResponse": {
        "properties": {
          "challenge_id": {
            "type": "string",
            "title": "Challenge Id"
          },
          "challenge": {
            "type": "string",
            "title": "Challenge",
            "description": "base64 nonce; sign it with the device key (ES256)"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "signing_input_hint": {
            "type": "string",
            "title": "Signing Input Hint",
            "default": "ES256 (raw r||s, base64url) over the ASCII bytes of `challenge`"
          }
        },
        "type": "object",
        "required": [
          "challenge_id",
          "challenge",
          "expires_at"
        ],
        "title": "StepUpChallengeResponse"
      },
...
      "UIComponent": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "portfolio_summary",
              "portfolio_positions",
              "attribution_bars",
              "cash_summary",
              "product_list",
              "product_detail",
              "product_comparison",
              "quote_table",
              "news_list",
              "research_list",
              "calendar_list",
              "simulation_chart",
              "fee_breakdown",
              "transaction_list",
              "statement_link",
              "accounts_list",
              "services_guide",
              "suitability_summary",
              "warning_banner",
              "form_spec",
              "citations",
              "escalation_offer",
              "escalation_card",
              "complaint_card",
              "order_receipt",
              "disclosure",
              "profile_update_offer"
            ],
            "title": "Type"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "payload"
        ],
        "title": "UIComponent",
        "description": "Every component carries ``source`` and ``as_of`` so the client can render\nstaleness and an auditor can trace it (docs/07-data-governance/01 §5)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}